# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Weekly Report API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/supervisor/staff": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Melihat siapa saja staff dari supervisor",
                "description": "Menampilkan list siapa saja staff-staff dari seorang supervisor atau vice supervisor",
                "operationId": "28a5d75252333aa9c16ebe4a44759ace",
                "responses": {
                    "200": {
                        "description": "Berhasil menampilkan list staff dari supervisor.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "John"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "Doe"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "null",
                                                        "nullable": true
                                                    },
                                                    "instagram": {
                                                        "type": "string",
                                                        "example": "null",
                                                        "nullable": true
                                                    },
                                                    "linkedin": {
                                                        "type": "string",
                                                        "example": "null",
                                                        "nullable": true
                                                    },
                                                    "batch_no": {
                                                        "type": "string",
                                                        "example": "null",
                                                        "nullable": true
                                                    },
                                                    "division": {
                                                        "type": "string",
                                                        "example": "null",
                                                        "nullable": true
                                                    },
                                                    "supervisor_id": {
                                                        "type": "integer",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "vice_supervisor_id": {
                                                        "type": "integer",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "CFlag": {
                                                        "type": "boolean",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "SFlag": {
                                                        "type": "boolean",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "StFlag": {
                                                        "type": "boolean",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "profile_picture": {
                                                        "type": "string",
                                                        "format": "uri",
                                                        "example": "http://example.com/profile.jpg"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "null",
                                                        "nullable": true
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "null",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Data Not Found. Token tidak valid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token tidak valid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/c-level/supervisor-staff/{divisionId}/list": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "C-Level melihat supervisor dan staff di divisi dia",
                "description": "Melihat semua staff dan supervisor di divisi yang dia naungi",
                "operationId": "ac09f80962ec83bd00797ddb2003ff23",
                "parameters": [
                    {
                        "name": "divisionId",
                        "in": "path",
                        "description": "ID divisi yang ingin dilihat",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data supervisor dan staff berhasil ditemukan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "division_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "division_name": {
                                            "type": "string",
                                            "example": "Web Development"
                                        },
                                        "team_members": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Elsie Lebsack"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "Head"
                                                    },
                                                    "profile_picture": {
                                                        "type": "string",
                                                        "format": "uri",
                                                        "example": "https://via.placeholder.com/640x480.png/0099ff?text=eum"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Token tidak valid atau tidak memiliki hak akses.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan internal server.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/division/staff-count": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Clevel melihat jumlah divisi dan jumlah staf",
                "description": "Endpoint ini mengembalikan jumlah divisi dan total staf yang ada di divisi-divisi tersebut untuk C-Level yang sedang login.",
                "operationId": "db9e0544ad758c35ddaf1dcde6ed07cd",
                "responses": {
                    "200": {
                        "description": "Jumlah divisi dan staf berhasil diambil.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "division_count": {
                                            "description": "Jumlah divisi yang dimiliki C-Level.",
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "total_staff_count": {
                                            "description": "Jumlah total staf yang ada di divisi-divisi C-Level.",
                                            "type": "integer",
                                            "example": 25
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Token tidak valid atau tidak ada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized: Token tidak valid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Terjadi kesalahan di server.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error: Terjadi kesalahan di server."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}